# investment-backend/openapi.yaml
openapi: 3.0.0
info:
  title: Investment Banking API
  version: 1.0.0
  description: API for managing investment banking clients and savings accounts.
servers:
  - url: http://localhost:8080/api
paths:
  /clients:
    get:
      summary: List all clients
      parameters:
        - in: query
          name: minInvestment
          required: false
          schema:
            type: number
            format: double
          description: Minimum total investment to filter clients
        - in: query
          name: maxInvestment
          required: false
          schema:
            type: number
            format: double
          description: Maximum total investment to filter clients
      responses:
        '200':
          description: A list of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
    post:
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        '201':
          description: Client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
...
components:
  schemas:
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        totalInvestment:
          type: number
          format: double
      required:
        - id
        - name
        - email
        - totalInvestment
    ClientCreate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        totalInvestment:
          type: number
          format: double
      required:
        - name
        - email
        - totalInvestment
...